Move `insert` function to the base module.
